Add fees to PDF report
