+++ pyproject.toml
@@ -65,10 +65,8 @@ max-line-length = 88
 # duplicate-code - unavoidable
 # too-many-* - not enforced for the sake of readability
 # too-few-* - same as too-many-*
-# others should be fixed
 disable = [
     "duplicate-code",
-    "fixme",
     "too-few-public-methods",
     "too-many-arguments",
     "too-many-branches",
